src/main/Servers/requirements.txt
argparse
pandas
numpy
matplotlib
scikit-learn
pyserial
micromlgen